test(files): separated smoke tests from the main
